complete the jwt token logic

+++ src/likes/likes.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, InternalServerErrorException } from '@nestjs/common';
 import { SupabaseService } from 'src/supabase/supabase.service';
 
 @Injectable()
@@ -12,14 +12,35 @@ export class LikesService {
             .select('*');
 
         if (error) {
-            throw new Error(error.message);
+            throw new InternalServerErrorException(error.message);
         }
 
         return data;
     }
 
-    async addLikes({ post_id, user_id }: { post_id: string; user_id: string }) {
+    async findLikesExist(post_id: string, user_id: string) {
         const { data, error } = await this.supabaseService
+            .getClient()
+            .from('likes')
+            .select('id')
+            .eq('post_id', post_id)
+            .eq('user_id', user_id)
+            .maybeSingle();
+
+        if (error) {
+            throw new InternalServerErrorException(error.message);
+        }
+
+        return data;
+    }
+
+    async addLikes(post_id: string, user_id: string) {
+        const likesExist = await this.findLikesExist(post_id, user_id);
+        if (likesExist) {
+            throw new BadRequestException("You already like this post");
+        }
+
+        const { error } = await this.supabaseService
             .getClient()
             .from('likes')
             .insert({
@@ -29,12 +50,19 @@ export class LikesService {
             });
 
         if (error) {
-            throw new Error(error.message);
+            throw new InternalServerErrorException(error.message);
         }
+
+        return { message: 'Successfully add likes' };
     }
 
-    async deleteLikes({ post_id, user_id }: { post_id: string; user_id: string }) {
-        const { data, error } = await this.supabaseService
+    async deleteLikes(post_id: string, user_id: string) {
+        const likesExist = await this.findLikesExist(post_id, user_id);
+        if (!likesExist) {
+            throw new BadRequestException("You are not like this post");
+        }
+
+        const { error } = await this.supabaseService
             .getClient()
             .from('likes')
             .delete()
@@ -42,7 +70,9 @@ export class LikesService {
             .eq('user_id', user_id);
 
         if (error) {
-            throw new Error(error.message);
+            throw new InternalServerErrorException(error.message);
         }
+
+        return { message: 'Successfully delete likes' };
     }
 }
